# domains/alert-managment/alert-service/src/main/resources/openapi.yml
openapi: 3.0.0
info:
  title: Alert Service API
  version: "1.0"
  description: REST API for creating and managing alerts in Cassandra
servers:
  - url: http://localhost:8081
paths:
  /api/alerts:
    get:
      summary: Get all alerts
      responses:
        "200":
          description: List of alerts
    post:
      summary: Create an alert
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Alert'
      responses:
        "201":
          description: Alert created
  /api/alerts/{id}:
    get:
      summary: Get alert by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Alert found
        "404":
          description: Not found
    put:
      summary: Update alert
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Alert'
      responses:
        "200":
          description: Alert updated
        "404":
          description: Not found
    delete:
      summary: Delete alert
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
components:
  schemas:
    Alert:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        message:
          type: string
        severity:
          type: string
        area:
          type: string
        timestamp:
          type: string
          format: date-time